make load data script better for future refactor
